Added logging functionality to pipelines
Incorporated logging to the assembly and batch assembly pipelines, a logging.json configuration file was added as well. Additionally, checks for existing output directory and input directory validation have been added to mitigate errors. The setup.py file was updated to reflect the addition of logging.json and the project status was updated to beta.