Add BBpath class and refactor mapping data handling
Introduced the BBpath class to encapsulate mapping file paths. Refactored the read mapping section to use the BBpath class and moved the genome extraction logic to a new function, assess_mapping_data.

Next steps are to amend this function to utilise coverage and mapping percentages to find the correct genome.